+++ tests/mdm/test_api_enrolled_devices_views.py
@@ -9,8 +9,9 @@ from django.utils.crypto import get_random_string
 from django.test import TestCase, override_settings
 from accounts.models import APIToken, User
 from zentral.contrib.inventory.models import MetaBusinessUnit, MachineTag, Tag
+from zentral.contrib.mdm.commands import SetAutoAdminPassword
 from zentral.contrib.mdm.commands.base import load_command
-from zentral.contrib.mdm.events import FileVaultPRKViewedEvent, RecoveryPasswordViewedEvent
+from zentral.contrib.mdm.events import AdminPasswordViewedEvent, FileVaultPRKViewedEvent, RecoveryPasswordViewedEvent
 from zentral.contrib.mdm.models import Platform
 from .utils import force_dep_enrollment_session
 
@@ -98,6 +99,9 @@ class APIViewsTestCase(TestCase):
              'next': None,
              'previous': None,
              'results': [{'activation_lock_manageable': None,
+                          'admin_guid': None,
+                          'admin_password_escrowed': False,
+                          'admin_shortname': None,
                           'apple_silicon': None,
                           'awaiting_configuration': None,
                           'blocked_at': None,
@@ -142,6 +146,9 @@ class APIViewsTestCase(TestCase):
              'next': None,
              'previous': None,
              'results': [{'activation_lock_manageable': None,
+                          'admin_guid': None,
+                          'admin_password_escrowed': False,
+                          'admin_shortname': None,
                           'apple_silicon': None,
                           'awaiting_configuration': None,
                           'blocked_at': None,
@@ -195,6 +202,9 @@ class APIViewsTestCase(TestCase):
              'next': None,
              'previous': None,
              'results': [{'activation_lock_manageable': None,
+                          'admin_guid': None,
+                          'admin_password_escrowed': False,
+                          'admin_shortname': None,
                           'apple_silicon': None,
                           'awaiting_configuration': None,
                           'blocked_at': None,
@@ -240,6 +250,12 @@ class APIViewsTestCase(TestCase):
         self.enrolled_device.set_bootstrap_token(b"un")
         self.enrolled_device.set_filevault_prk("deux")
         self.enrolled_device.set_recovery_password("trois")
+        self.enrolled_device.device_information = {
+            "AutoSetupAdminAccounts": [
+                {"GUID": "yolo", "shortName": "fomo"}
+            ]
+        }
+        self.enrolled_device.set_admin_password("quatre")
         self.enrolled_device.save()
         self.set_permissions("mdm.view_enrolleddevice")
         response = self.get(reverse("mdm_api:enrolled_devices"))
@@ -250,6 +266,9 @@ class APIViewsTestCase(TestCase):
              'next': None,
              'previous': None,
              'results': [{'activation_lock_manageable': None,
+                          'admin_guid': "yolo",
+                          'admin_password_escrowed': True,
+                          'admin_shortname': "fomo",
                           'apple_silicon': None,
                           'awaiting_configuration': None,
                           'blocked_at': None,
@@ -323,6 +342,9 @@ class APIViewsTestCase(TestCase):
              'next': None,
              'previous': None,
              'results': [{'activation_lock_manageable': None,
+                          'admin_guid': None,
+                          'admin_password_escrowed': False,
+                          'admin_shortname': None,
                           'apple_silicon': None,
                           'awaiting_configuration': None,
                           'blocked_at': None,
@@ -397,6 +419,9 @@ class APIViewsTestCase(TestCase):
         self.assertEqual(
             response.json(),
             {'activation_lock_manageable': None,
+             'admin_guid': None,
+             'admin_password_escrowed': False,
+             'admin_shortname': None,
              'apple_silicon': None,
              'awaiting_configuration': None,
              'blocked_at': self.enrolled_device.blocked_at.isoformat(),
@@ -456,6 +481,9 @@ class APIViewsTestCase(TestCase):
         self.assertEqual(
             response.json(),
             {'activation_lock_manageable': None,
+             'admin_guid': None,
+             'admin_password_escrowed': False,
+             'admin_shortname': None,
              'apple_silicon': None,
              'awaiting_configuration': None,
              'blocked_at': None,
@@ -974,3 +1002,114 @@ class APIViewsTestCase(TestCase):
         event = post_event.call_args_list[0].args[0]
         self.assertIsInstance(event, RecoveryPasswordViewedEvent)
         self.assertEqual(event.metadata.machine_serial_number, self.enrolled_device.serial_number)
+
+    # enrolled device admin password
+
+    def test_enrolled_device_admin_password_unauthorized(self):
+        response = self.get(reverse("mdm_api:enrolled_device_admin_password", args=(self.enrolled_device.pk,)),
+                            include_token=False)
+        self.assertEqual(response.status_code, 401)
+
+    def test_enrolled_device_admin_password_permission_denied(self):
+        response = self.get(reverse("mdm_api:enrolled_device_admin_password", args=(self.enrolled_device.pk,)))
+        self.assertEqual(response.status_code, 403)
+
+    def test_enrolled_device_admin_password_login_permission_denied(self):
+        self.login()
+        response = self.get(reverse("mdm_api:enrolled_device_admin_password", args=(self.enrolled_device.pk,)),
+                            include_token=False)
+        self.assertEqual(response.status_code, 403)
+
+    @patch("zentral.core.queues.backends.kombu.EventQueues.post_event")
+    def test_enrolled_device_admin_password_null(self, post_event):
+        self.set_permissions("mdm.view_admin_password")
+        response = self.get(reverse("mdm_api:enrolled_device_admin_password", args=(self.enrolled_device.pk,)))
+        self.assertEqual(
+            response.json(),
+            {"id": self.enrolled_device.pk,
+             "serial_number": self.enrolled_device.serial_number,
+             "admin_password": None}
+        )
+        post_event.assert_not_called()
+
+    @patch("zentral.core.queues.backends.kombu.EventQueues.post_event")
+    def test_enrolled_device_admin_password(self, post_event):
+        self.assertEqual(self.dep_enrollment_session.dep_enrollment.admin_password_rotation_delay, 60)
+        self.enrolled_device.set_admin_password("123456")
+        self.enrolled_device.device_information = {
+            "AutoSetupAdminAccounts": [
+                {"GUID": "yolo", "shortName": "fomo"}
+            ]
+        }
+        self.enrolled_device.save()
+        self.set_permissions("mdm.view_admin_password")
+        response = self.get(reverse("mdm_api:enrolled_device_admin_password", args=(self.enrolled_device.pk,)))
+        self.assertEqual(
+            response.json(),
+            {"id": self.enrolled_device.pk,
+             "serial_number": self.enrolled_device.serial_number,
+             "admin_password": "123456"}
+        )
+        self.assertEqual(len(post_event.call_args_list), 1)
+        event = post_event.call_args_list[0].args[0]
+        self.assertIsInstance(event, AdminPasswordViewedEvent)
+        self.assertEqual(event.metadata.machine_serial_number, self.enrolled_device.serial_number)
+        command_qs = self.enrolled_device.commands.all()
+        self.assertEqual(command_qs.count(), 1)
+        command = load_command(command_qs.first())
+        self.assertIsInstance(command, SetAutoAdminPassword)
+
+    @patch("zentral.core.queues.backends.kombu.EventQueues.post_event")
+    def test_enrolled_device_admin_password_login(self, post_event):
+        self.assertEqual(self.dep_enrollment_session.dep_enrollment.admin_password_rotation_delay, 60)
+        self.enrolled_device.set_admin_password("123456")
+        self.enrolled_device.device_information = {
+            "AutoSetupAdminAccounts": [
+                {"GUID": "yolo", "shortName": "fomo"}
+            ]
+        }
+        self.enrolled_device.save()
+        self.login("mdm.view_admin_password")
+        response = self.get(reverse("mdm_api:enrolled_device_admin_password", args=(self.enrolled_device.pk,)),
+                            include_token=False)
+        self.assertEqual(
+            response.json(),
+            {"id": self.enrolled_device.pk,
+             "serial_number": self.enrolled_device.serial_number,
+             "admin_password": "123456"}
+        )
+        self.assertEqual(len(post_event.call_args_list), 1)
+        event = post_event.call_args_list[0].args[0]
+        self.assertIsInstance(event, AdminPasswordViewedEvent)
+        self.assertEqual(event.metadata.machine_serial_number, self.enrolled_device.serial_number)
+        command_qs = self.enrolled_device.commands.all()
+        self.assertEqual(command_qs.count(), 1)
+        command = load_command(command_qs.first())
+        self.assertIsInstance(command, SetAutoAdminPassword)
+
+    @patch("zentral.core.queues.backends.kombu.EventQueues.post_event")
+    def test_enrolled_device_admin_password_login_no_auto_rotation(self, post_event):
+        self.dep_enrollment_session.dep_enrollment.admin_password_rotation_delay = 0  # no auto rotation
+        self.dep_enrollment_session.dep_enrollment.save()
+        self.enrolled_device.set_admin_password("123456")
+        self.enrolled_device.device_information = {
+            "AutoSetupAdminAccounts": [
+                {"GUID": "yolo", "shortName": "fomo"}
+            ]
+        }
+        self.enrolled_device.save()
+        self.login("mdm.view_admin_password")
+        response = self.get(reverse("mdm_api:enrolled_device_admin_password", args=(self.enrolled_device.pk,)),
+                            include_token=False)
+        self.assertEqual(
+            response.json(),
+            {"id": self.enrolled_device.pk,
+             "serial_number": self.enrolled_device.serial_number,
+             "admin_password": "123456"}
+        )
+        self.assertEqual(len(post_event.call_args_list), 1)
+        event = post_event.call_args_list[0].args[0]
+        self.assertIsInstance(event, AdminPasswordViewedEvent)
+        self.assertEqual(event.metadata.machine_serial_number, self.enrolled_device.serial_number)
+        command_qs = self.enrolled_device.commands.all()
+        self.assertEqual(command_qs.count(), 0)