+++ tests/mdm/test_models.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 import uuid
 from django.test import TestCase
 from django.utils.crypto import get_random_string
@@ -85,6 +86,8 @@ class TestMDMModels(TestCase):
         enrolled_device.set_recovery_password(recovery_password)
         pending_firmware_password = get_random_string(12)
         enrolled_device.set_pending_firmware_password(pending_firmware_password)
+        admin_password = get_random_string(12)
+        enrolled_device.set_admin_password(admin_password)
         enrolled_device.rewrap_secrets()
         self.assertEqual(enrolled_device.get_bootstrap_token(), bootstrap_token)
         self.assertEqual(enrolled_device.get_unlock_token(), unlock_token)
@@ -92,6 +95,9 @@ class TestMDMModels(TestCase):
         self.assertEqual(enrolled_device.get_filevault_prk(), filevault_prk)
         self.assertEqual(enrolled_device.get_recovery_password(), recovery_password)
         self.assertEqual(enrolled_device.get_pending_firmware_password(), pending_firmware_password)
+        self.assertEqual(enrolled_device.get_admin_password(), admin_password)
+
+    # urlsafe serial number
 
     def test_enrolled_device_get_urlsafe_serial_number(self):
         enrolled_device = EnrolledDevice.objects.create(
@@ -100,3 +106,59 @@ class TestMDMModels(TestCase):
             push_certificate=force_push_certificate(),
         )
         self.assertEqual("0123456789", enrolled_device.get_urlsafe_serial_number())
+
+    # current enrollment / enrollment session
+
+    def test_enrolled_device_blocked_no_current_enrollment_session(self):
+        enrolled_device = EnrolledDevice.objects.create(
+            udid=uuid.uuid4(),
+            serial_number="0123456789",
+            push_certificate=force_push_certificate(),
+            blocked_at=datetime.utcnow(),
+        )
+        self.assertIsNone(enrolled_device.current_enrollment_session)
+        self.assertIsNone(enrolled_device.current_enrollment)
+
+    def test_enrolled_device_no_current_enrollment_session(self):
+        enrolled_device = EnrolledDevice.objects.create(
+            udid=uuid.uuid4(),
+            serial_number="0123456789",
+            push_certificate=force_push_certificate(),
+        )
+        self.assertIsNone(enrolled_device.current_enrollment_session)
+        self.assertIsNone(enrolled_device.current_enrollment)
+
+    # auto admin
+
+    def test_enrolled_device_auto_admin_info(self):
+        enrolled_device = EnrolledDevice.objects.create(
+            udid=uuid.uuid4(),
+            serial_number="0123456789",
+            push_certificate=force_push_certificate(),
+            device_information={
+                "AutoSetupAdminAccounts": [
+                    {"GUID": "yolo", "shortName": "fomo"}
+                ]
+            }
+        )
+        self.assertEqual(enrolled_device.admin_guid, "yolo")
+        self.assertEqual(enrolled_device.admin_shortname, "fomo")
+        self.assertFalse(enrolled_device.admin_password_escrowed)
+        enrolled_device.set_admin_password("pwd")
+        self.assertTrue(enrolled_device.admin_password_escrowed)
+
+    def test_enrolled_device_no_auto_admin_info(self):
+        enrolled_device = EnrolledDevice.objects.create(
+            udid=uuid.uuid4(),
+            serial_number="0123456789",
+            push_certificate=force_push_certificate(),
+            device_information={
+                "AutoSetupAdminAccounts": [
+                    {"GUID": "yolo", "shortName": "fomo"},
+                    {"GUID": "yolo2", "shortName": "fomo2"}  # Zentral only manages 1 auto admin
+                ]
+            }
+        )
+        self.assertIsNone(enrolled_device.admin_guid)
+        self.assertIsNone(enrolled_device.admin_shortname)
+        self.assertFalse(enrolled_device.admin_password_escrowed)